+++ lib/browser/browser.js
@@ -27,7 +27,6 @@ var token             = require('../token');
 var TokenManager      = require('../TokenManager');
 var tx                = require('../tx');
 var util              = require('../util');
-var sdkCrypto         = require('../crypto');
 
 function OktaAuthBuilder(args) {
   var sdk = this;
@@ -96,9 +95,7 @@ function OktaAuthBuilder(args) {
 
   sdk.pkce = {
     generateVerifier: pkce.generateVerifier,
-    computeChallenge: sdkCrypto.computeChallenge,
-    getWithRedirect: util.bind(pkce.getWithRedirect, null, sdk),
-    parseFromUrl: util.bind(pkce.parseFromUrl, null, sdk),
+    computeChallenge: pkce.computeChallenge,
     exchangeForToken: util.bind(pkce.exchangeForToken, null, sdk)
   };
