+++ lib/crypto.js
@@ -12,17 +12,6 @@
 
 var util = require('./util');
 
-/* global Uint8Array */
-function computeChallenge(str) {  
-  var buffer = new TextEncoder('utf-8').encode(str);
-  return crypto.subtle.digest('SHA-256', buffer).then(function(arrayBuffer) {
-    var hash = String.fromCharCode.apply(null, new Uint8Array(arrayBuffer));
-    var b64 = btoa(hash); // standard base64 encode
-    var b64u = util.base64ToBase64Url(b64); // use url-safe base64 variant
-    return b64u;
-  });
-}
-
 function verifyToken(idToken, key) {
   key = util.clone(key);
 
@@ -62,6 +51,5 @@ function verifyToken(idToken, key) {
 }
 
 module.exports = {
-  computeChallenge: computeChallenge,
   verifyToken: verifyToken
 };
