+++ lib/pkce.js
@@ -20,31 +20,27 @@ var token      = require('./token');
 
 // Code verifier: Random URL-safe string with a minimum length of 43 characters.
 // Code challenge: Base64 URL-encoded SHA-256 hash of the code verifier.
-
 var MIN_VERIFIER_LENGTH = 43;
 function generateVerifier(prefix) {
   var str = prefix;
   if (str.length < MIN_VERIFIER_LENGTH) {
-    str = str + util.genRandomString(MIN_VERIFIER_LENGTH);
+    str = str + util.genRandomString(MIN_VERIFIER_LENGTH - str.length);
   }
   return encodeURIComponent(str);
 }
 
-function getWithRedirect(sdk, oauthOptions, options) {
-  oauthOptions = oauthOptions || {};
-  oauthOptions.responseType = oauthOptions.responseType || 'code';
-  oauthOptions.responseMode = oauthOptions.responseMode || 'fragment';
-
-  return token.getWithRedirect(sdk, oauthOptions, options);
+/* global Uint8Array */
+function computeChallenge(str) {  
+  var buffer = new TextEncoder('utf-8').encode(str);
+  return crypto.subtle.digest('SHA-256', buffer).then(function(arrayBuffer) {
+    var hash = String.fromCharCode.apply(null, new Uint8Array(arrayBuffer));
+    var b64 = btoa(hash); // standard base64 encode
+    var b64u = util.base64ToBase64Url(b64); // use url-safe base64 variant
+    return b64u;
+  });
 }
 
-function parseFromUrl(sdk, url) {
-  return sdk.token.parseFromUrl(url)
-    .then(function(res) {
-      return res.authorizationCode;
-    });
-}
-
+// for /v1/token endpoint
 function setDefaultOptions(sdk, options) {
   options = util.clone(options) || {};
   var defaults = {
@@ -56,21 +52,15 @@ function setDefaultOptions(sdk, options) {
   return defaults;
 }
 
-function validateOptions(oauthParams) {
+function validateOptions(oauthOptions) {
   // Quick validation
-  if (!oauthParams.clientId) {
+  if (!oauthOptions.clientId) {
     throw new AuthSdkError('A clientId must be specified in the OktaAuth constructor to get a token');
   }
-
-  if (util.isString(oauthParams.responseType) && oauthParams.responseType.indexOf(' ') !== -1) {
-    throw new AuthSdkError('Multiple OAuth responseTypes must be defined as an array');
-  }
-
 }
 
-// for /v1/token endpoint
 function getQueryParams(options) {
-  // Convert our options to OAuth params
+  // Convert options to OAuth params
   var params = util.removeNils({
     'client_id': options.clientId,
     'redirect_uri': options.redirectUri,
@@ -79,14 +69,10 @@ function getQueryParams(options) {
     'code_verifier': options.codeVerifier
   });
 
-  if (Array.isArray(params['response_type'])) {
-    params['response_type'] = params['response_type'].join(' ');
-  }
-
   return params;
 }
 
-// exchange authoriza code for an access token
+// exchange authorization code for an access token
 function exchangeForToken(sdk, oauthOptions, options) {
   oauthOptions = setDefaultOptions(sdk, oauthOptions || {});
   options = options || {};
@@ -103,20 +89,17 @@ function exchangeForToken(sdk, oauthOptions, options) {
     withCredentials: false
   })
   .then(function(res) {
-
     if (!oauthOptions.responseType) {
       oauthOptions.responseType = ['id_token', 'token'];
     }
-
+    // scopes were passed in original /authorize call and are returned to us in this response 
     oauthOptions.scopes = res.scope.split(' ');
-
     return token.handleOAuthResponse(sdk, oauthOptions, res, urls);
   });
 }
 
 module.exports = {
-  exchangeForToken: exchangeForToken,
-  getWithRedirect: getWithRedirect,
-  parseFromUrl: parseFromUrl,
-  generateVerifier: generateVerifier
+  generateVerifier: generateVerifier,
+  computeChallenge: computeChallenge,
+  exchangeForToken: exchangeForToken
 };
