+++ lib/util.js
@@ -12,14 +12,6 @@
 /* eslint-env es6 */
 var util = module.exports;
 
-util.bufferToHexString = function (buffer) {
-  return Array.prototype.map.call(new Uint8Array(buffer), function(x) {
-    var val = ('00' + x.toString(16)).slice(-2); //.join('');
-    // console.log('x', x, val);
-    return val;
-  }).join('');
-};
-
 // converts a standard base64-encoded string to a "url/filename safe" variant
 util.base64ToBase64Url = function(b64) {
   return b64.replace(/\+/g, '-').replace(/\//g, '_').replace(/=+$/, '');
